Add Y2Partitioner::Actions::Controllers::Base
